+++ examples/optimize_multidiag.py
@@ -56,7 +56,8 @@ if __name__ == "__main__":
                         Path("/home/weis/Desktop/HDR/7LX1_2/hdr/11_3662.npy"),
                         Path("/home/weis/Desktop/HDR/7LX1_2/hdr/12_3840.npy"),
                         Path("/home/weis/Desktop/HDR/7LX1_2/hdr/13_4018.npy"))
-    def_images = tuple(np.load(img) for img in def_images_paths)
+    def_images = tuple(np.load(img).astype(np.float64)
+                       for img in def_images_paths)
 
     # Effort measured during the test
     efforts_file = Path("/home/weis/Desktop/HDR/7LX1_2/effort_2.csv")
@@ -86,11 +87,11 @@ if __name__ == "__main__":
 
     # Optimize an all the images at once
     optimize_diagonals(lib_path,
-                       ref_img,
-                       density_base,
-                       gauss_fit,
-                       peaks,
-                       x0,
+                       ref_img.astype(np.float64),
+                       density_base.astype(np.float64),
+                       gauss_fit.astype(np.float64),
+                       peaks.astype(np.float64),
+                       x0.astype(np.float64),
                        def_images,
                        efforts_x,
                        order_coeffs,
